+++ lib/wasmboy/worker/instantiate.js
@@ -1,10 +1,11 @@
 // Function to instantiate our wasm module
 import { postMessage } from '../../worker/workerapi';
 import { isInBrowser, readBase64Buffer } from '../../worker/util';
+import * as WasmBoyCore from '../../../dist/core/core.esm';
 import importObject from './importobject';
 
 // Function to instantiate our wasm and respond back
-export const instantiateWasm = async wasmModuleUrl => {
+export const instantiateWasmCore = async wasmModuleUrl => {
   let response = undefined;
 
   if (isInBrowser) {
@@ -49,4 +50,14 @@ const wasmNodeInstantiate = async wasmModuleUrl => {
       performanceTimestamp: () => {}
     }
   });
+};
+
+export const instantiateTsCore = async () => {
+  const response = {
+    wasmInstance: {
+      exports: WasmBoyCore
+    },
+    wasmByteMemory: WasmBoyCore.memory.wasmByteMemory
+  };
+  return response;
 };
