+++ lib/wasmboy/worker/wasmboy.worker.js
@@ -15,8 +15,7 @@ import { controllerWorkerOnMessage } from './controller/onmessage';
 import { memoryWorkerOnMessage } from './memory/onmessage';
 
 // Working the emalator
-import { instantiateWasm } from './instantiate';
-import * as WasmBoyCore from '../../../dist/core.esm';
+import { instantiateWasmCore, instantiateTsCore } from './instantiate';
 import { update } from './update';
 
 // Timestamps
@@ -100,21 +99,12 @@ libWorker = {
       case WORKER_MESSAGE_TYPE.INSTANTIATE_WASM: {
         // TODO: Replace this conditionallyy
         const instantiateTask = async () => {
-          const useWasm = true;
-          console.log('Using Wasm?', useWasm);
-          if (useWasm) {
-            const response = await instantiateWasm(eventData.message.wasmModuleUrl);
-            libWorker.wasmInstance = response.wasmInstance;
-            libWorker.wasmByteMemory = response.wasmByteMemory;
-            postMessage(getSmartWorkerMessage(undefined, eventData.messageId));
-            return;
-          }
+          let response;
+          response = await instantiateWasmCore(eventData.message.wasmModuleUrl);
+          response = await instantiateTsCore();
 
-          console.log('core', WasmBoyCore);
-          libWorker.wasmInstance = {
-            exports: WasmBoyCore
-          };
-          libWorker.wasmByteMemory = WasmBoyCore.memory.wasmByteMemory;
+          libWorker.wasmInstance = response.wasmInstance;
+          libWorker.wasmByteMemory = response.wasmByteMemory;
           postMessage(getSmartWorkerMessage(undefined, eventData.messageId));
           return;
         };
