+++ lib/worker/instantiate.js
@@ -5,11 +5,13 @@ import { WasmBoyController } from '../controller/controller';
 import { WasmBoyMemory } from '../memory/memory';
 
 // Import our worker urls
-import wasmboyLibWorkerUrl from '../../dist/wasmboy/worker/wasmboy.worker.js';
-import wasmboyGraphicsWorkerUrl from '../../dist/graphics/worker/graphics.worker.js';
-import wasmboyAudioWorkerUrl from '../../dist/audio/worker/audio.worker.js';
-import wasmboyControllerWorkerUrl from '../../dist/controller/worker/controller.worker.js';
-import wasmboyMemoryWorkerUrl from '../../dist/memory/worker/memory.worker.js';
+// Only one of the wasmboylib imports will be used. Rollup/Babel handles which one
+import wasmboyLibWasmWorkerUrl from '../../dist/worker/wasmboy.wasm.worker.js';
+import wasmboyLibTsWorkerUrl from '../../dist/worker/wasmboy.ts.worker.js';
+import wasmboyGraphicsWorkerUrl from '../../dist/worker/graphics.worker.js';
+import wasmboyAudioWorkerUrl from '../../dist/worker/audio.worker.js';
+import wasmboyControllerWorkerUrl from '../../dist/worker/controller.worker.js';
+import wasmboyMemoryWorkerUrl from '../../dist/worker/memory.worker.js';
 
 // Import our Smart Worker Interface
 import { SmartWorker } from './smartworker';
@@ -21,7 +23,11 @@ ROLLUP_REPLACE_NODE*/
 
 export const instantiateWorkers = async () => {
   // Create our workers
-  const libWorker = new SmartWorker(wasmboyLibWorkerUrl, WORKER_ID.LIB);
+  let libWorkerUrl;
+  libWorkerUrl = wasmboyLibWasmWorkerUrl;
+  libWorkerUrl = wasmboyLibTsWorkerUrl;
+
+  const libWorker = new SmartWorker(libWorkerUrl, WORKER_ID.LIB);
   const graphicsWorker = new SmartWorker(wasmboyGraphicsWorkerUrl, WORKER_ID.GRAPHICS);
   const audioWorker = new SmartWorker(wasmboyAudioWorkerUrl, WORKER_ID.AUDIO);
   const controllerWorker = new SmartWorker(wasmboyControllerWorkerUrl, WORKER_ID.CONTROLLER);
