+++ package.json
@@ -16,8 +16,9 @@
   "module": "dist/wasmboy.esm.js",
   "browser": "dist/wasmboy.umd.js",
   "scripts": {
-    "start": "npm run core:build && npm run lib:build && npx concurrently --kill-others --names \"DEBUGGER,CORE,LIB\" -c \"bgBlue.bold,bgMagenta.bold,bgGreen.bold\" \"npm run debugger:watch\" \"npm run core:watch\" \"npm run lib:watch\"",
-    "start:ts": "npm run lib:build:ts && npx concurrently --kill-others --names \"DEBUGGER,LIBANDCORETS\" -c \"bgBlue.bold,bgGreen.bold\" \"npm run debugger:watch\" \"npm run lib:watch:ts\"",
+    "prepare": "npm run core:build && npm run lib:build && npm run lib:build:ts",
+    "start": "npx concurrently --kill-others --names \"DEBUGGER,CORE,LIB\" -c \"bgBlue.bold,bgMagenta.bold,bgGreen.bold\" \"npm run debugger:watch\" \"npm run core:watch\" \"npm run lib:watch\"",
+    "start:ts": "npx concurrently --kill-others --names \"DEBUGGER,LIBANDCORETS\" -c \"bgBlue.bold,bgGreen.bold\" \"npm run debugger:watch\" \"npm run lib:watch:ts\"",
     "dev": "npm run start",
     "watch": "npm run start",
     "dev:ts": "npm run start:ts",
